Prepare the transition to protected mode
To prepare the transition to protected mode we started to remove the usage of
BIOS services in the kernel. It is not so easy to do it properly. So we start
the modification in the bootloader to load the kernel. The goal is to put in
kernel.asm the same thing that we have in stage2.asm (that is not used any more)
but without any BIOS interrupt.